moved game mod to shared mod
